Fix setup and other things
