+++ packages/react-app/src/index.tsx
@@ -15,13 +15,12 @@ import { App } from './App';
 import { IpfsContextProvider } from './ipfs/IpfsContextProvider';
 
 // Change this to your own Infura project id: https://infura.io/register
-const INFURA_PROJECT_ID = 'defba93b47f748f09fcead8282b9e58e' as const;
 const config: Config = {
-  readOnlyChainId: Goerli.chainId,
+  // readOnlyChainId: Mainnet.chainId,
   readOnlyUrls: {
-    [Mainnet.chainId]: `https://mainnet.infura.io/v3/${INFURA_PROJECT_ID}`,
-    [Goerli.chainId]: `https://goerli.infura.io/v3/${INFURA_PROJECT_ID}`,
-    [Localhost.chainId]: 'http://127.0.0.1:8545',
+    [Mainnet.chainId]: `https://mainnet.infura.io/v3/${process.env.REACT_APP_INFURA_PROJECT_ID}`,
+    [Goerli.chainId]: `https://goerli.infura.io/v3/${process.env.REACT_APP_INFURA_PROJECT_ID}`,
+    // [Localhost.chainId]: 'http://127.0.0.1:8545',
   },
 };
